# docs/openapi.yaml
openapi: 3.0.3
info:
  title: VideoCMS REST API
  description: >
    API REST para a plataforma VideoCMS / VMS de descoberta, gerenciamento,
    diagnóstico e transmissão ao vivo de câmeras IP e feeds RTSP em rede local.
  version: 1.0.0
servers:
  - url: http://localhost:15000/api
    description: Servidor Local VideoCMS
paths:
...
  /cameras:
    get:
      summary: Lista catálogo de câmeras
      description: Retorna lista de câmeras com suporte a filtros por status, grupo, tag, busca e paginação.
      parameters:
        - name: search
          in: query
          description: Busca por nome, host, fabricante ou modelo
          schema:
            type: string
        - name: status
          in: query
          description: Filtro por status (online, offline, auth_required, error)
          schema:
            type: string
        - name: group
          in: query
          description: Filtro por nome do grupo
          schema:
            type: string
        - name: tag
          in: query
          description: Filtro por tag
          schema:
            type: string
        - name: limit
          in: query
          description: Quantidade máxima de registros
          schema:
            type: integer
        - name: offset
          in: query
          description: Deslocamento para paginação
          schema:
            type: integer
      responses:
        '200':
          description: Lista de câmeras cadastradas
          headers:
            X-Total-Count:
              schema:
                type: integer
              description: Total de registros encontrados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Camera'
    post:
      summary: Cadastra nova câmera IP
      description: Cadastra manualmente uma câmera no catálogo com criptografia AES-256-GCM em repouso.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CameraInput'
      responses:
        '201':
          description: Câmera criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Camera'
        '400':
          $ref: '#/components/responses/400Error'
  /cameras/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Detalhes de uma câmera
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Camera'
        '404':
          $ref: '#/components/responses/404Error'
    put:
      summary: Atualiza configurações da câmera
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CameraInput'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Camera'
        '404':
          $ref: '#/components/responses/404Error'
    delete:
      summary: Remove câmera do catálogo
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/404Error'
  /cameras/{id}/update-ip:
    post:
      summary: Atualiza endereço IP da câmera (DHCP)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - host
              properties:
                host:
                  type: string
                  example: 192.168.1.105
                port:
                  type: integer
                  example: 80
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
  /cameras/{id}/test:
    post:
      summary: Testa conectividade e autenticação
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  error:
                    type: string
  /cameras/{id}/capabilities:
    get:
      summary: Consulta recursos e perfis de vídeo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CameraCapabilities'
  /cameras/{id}/diagnostics:
    get:
      summary: Executa diagnóstico técnico em 10 etapas
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnosticReport'
  /cameras/{id}/snapshot:
    get:
      summary: Captura imagem JPEG instantânea
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Imagem JPEG
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
...
  /discovery:
    get:
      summary: Lista tarefas recentes de descoberta
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiscoveryJob'
    post:
      summary: Inicia nova tarefa de descoberta (WS-Discovery + CIDR)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                interface_name:
                  type: string
                  example: eth0
                cidr:
                  type: string
                  example: 192.168.1.0/24
      responses:
        '202':
          description: Tarefa iniciada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveryJob'
  /discovery/{id}:
    get:
      summary: Detalhes e dispositivos de uma tarefa de descoberta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveryJob'
  /discovery/{id}/cancel:
    post:
      summary: Cancela tarefa de descoberta em andamento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /discovery/{id}/devices/{deviceId}/probe:
    post:
      summary: Sonda dispositivo descoberto com credenciais
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveryResult'
  /discovery/{id}/devices/{deviceId}/add:
    post:
      summary: Importa dispositivo descoberto para o catálogo de câmeras
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                username:
                  type: string
                password:
                  type: string
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Camera'
  /layouts:
    get:
      summary: Lista layouts de mosaico salvos
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Layout'
    post:
      summary: Salva novo layout de mosaico
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LayoutInput'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Layout'
  /layouts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Detalhes de um layout
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Layout'
    put:
      summary: Atualiza layout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LayoutInput'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Layout'
    delete:
      summary: Remove layout
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /groups:
    get:
      summary: Lista grupos de câmeras
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
    post:
      summary: Cria novo grupo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
...
components:
  responses:
    400Error:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    404Error:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    CameraCapabilities:
      type: object
      properties:
        onvif:
          type: boolean
        rtsp:
          type: boolean
        ptz:
          type: boolean
        audio:
          type: boolean
        snapshot:
          type: boolean
        events:
          type: boolean
        main_stream:
          type: boolean
        sub_stream:
          type: boolean
        profiles:
          type: array
          items:
            $ref: '#/components/schemas/StreamProfile'
    Camera:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        host:
          type: string
        port:
          type: integer
        rtsp_port:
          type: integer
        manufacturer:
          type: string
        model:
          type: string
        serial_number:
          type: string
        mac_address:
          type: string
        firmware_version:
          type: string
        onvif_url:
          type: string
        rtsp_path:
          type: string
        substream_path:
          type: string
        snapshot_path:
          type: string
        username:
          type: string
        has_password:
          type: boolean
        preferred_profile:
          type: string
        preferred_transport:
          type: string
        codec:
          type: string
        capabilities:
          $ref: '#/components/schemas/CameraCapabilities'
        enabled:
          type: boolean
        status:
          type: string
          enum: [unknown, online, offline, auth_required, unsupported, error]
        status_message:
          type: string
        last_seen_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
        groups:
          type: array
          items:
            type: string
    CameraInput:
      type: object
      required:
        - name
        - host
      properties:
        name:
          type: string
        host:
          type: string
        port:
          type: integer
          default: 80
        rtsp_port:
          type: integer
          default: 554
        manufacturer:
          type: string
        model:
          type: string
        username:
          type: string
        password:
          type: string
        preferred_transport:
          type: string
          enum: [tcp, udp]
          default: tcp
        onvif_url:
          type: string
        rtsp_path:
          type: string
        substream_path:
          type: string
        enabled:
          type: boolean
          default: true
        tags:
          type: array
          items:
            type: string
        groups:
          type: array
          items:
            type: string
...
    DiagnosticReport:
      type: object
      properties:
        camera_id:
          type: string
        camera_name:
          type: string
        host:
          type: string
        stages:
          type: array
          items:
            $ref: '#/components/schemas/DiagnosticStage'
        passed:
          type: boolean
        summary:
          type: string
        capabilities:
          $ref: '#/components/schemas/CameraCapabilities'
        tested_at:
          type: string
          format: date-time
...
    DiscoveryJob:
      type: object
      properties:
        id:
          type: string
        interface_name:
          type: string
        cidr:
          type: string
        status:
          type: string
          enum: [queued, running, completed, cancelled, failed]
        progress:
          type: integer
        total_hosts:
          type: integer
        scanned_hosts:
          type: integer
        found_devices:
          type: integer
        error_message:
          type: string
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        results:
          type: array
          items:
            $ref: '#/components/schemas/DiscoveryResult'
...